2. Cercador/requirements.txt
tensorflow
anvil-uplink
requests